Agregada la entidad Mail y MailAccount con los primeros mapeos

+++ glimpse/DataAccessLayer/Entities/Address.cs
@@ -8,7 +8,7 @@ namespace Glimpse.DataAccessLayer.Entities
 {
     public class Address
     {
-        public virtual UInt64 Id { get; set; }
+        public virtual int Id { get; set; }
         public virtual String MailAddress { get; set; }
         public virtual String Name { get; set; }
     }
