+++ glimpse/DataAccessLayer/Entities/Mail.cs
@@ -8,7 +8,7 @@ namespace Glimpse.DataAccessLayer.Entities
 {
     public class Mail
     {
-        public virtual UInt64 Id { get; set; }
+        public virtual int Id { get; set; }
         //public virtual MailAccount IdMailAccount { get; set; }
         public virtual Address From { get; set; }
         public virtual DateTime Date { get; set; }
